externalize env test
so convex can import it
